requirements.txt
certifi==2019.9.11
chardet==3.0.4
construct==2.5.3
defusedxml==0.6.0
Django==2.2.6
django-allauth==0.40.0
django-cors-headers==3.1.1
django-embed-video==1.3
django-haystack==2.8.1
django-rest-auth==0.9.5
djangorestframework==3.10.3
future==0.18.1
idna==2.8
oauthlib==3.1.0
pefile==2019.4.18
Pillow==6.2.1
PyJWT==1.7.1
pysolr==3.8.1
python-ptrace==0.9.4
python3-openid==3.1.0
pytz==2019.3
redis==3.3.11
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sorl-thumbnail==12.5.0
sqlparse==0.3.0
urllib3==1.25.6
whitenoise==4.1.4
